% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hodgson.R
\name{hodgson}
\alias{hodgson}
\title{Calculate CSR Scores and Assign a CSR Strategy Using the Hodgson et al. (1999) Model}
\usage{
hodgson(data, calcLDMC = FALSE, calcSLA = FALSE)
}
\arguments{
\item{data}{A data frame containing species names and trait data. Must include (LFW and LA optional instead of LDMC and SLA respectively):
\describe{
  \item{CH}{Canopy height (mm)}
  \item{LDMC}{Leaf dry matter content (\%)}
  \item{FP}{Flowering period (months)}
  \item{LS}{Lateral spread (categorical scale: 1–6, based on Hodgson et al. 1999)}
  \item{LDW}{Leaf dry weight (mg)}
  \item{SLA}{Specific leaf area (mm²/mg)}
  \item{LFW}{Leaf fresh weight (mg)}
  \item{LA}{Leaf area (mm²)}
}}

\item{calcLDMC}{Logical. Whether to calculate LDMC from LDW and LFW. Default is FALSE.}

\item{calcSLA}{Logical. Whether to calculate SLA from LA and LDW. Default is FALSE.}
}
\value{
A data frame with CSR scores (-2.5 to 2.5), CSR percentages (C%, S%, R%), and assigned strategy classification.
}
\description{
This function calculates and assigns CSR strategy types to plant species
based on the Hodgson et al. (1999) model, as well as performing
the required trait transformations for the model.
}
\examples{
# hodgson(exampleData)
}
