% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphoPhys.R
\name{morphoPhys}
\alias{morphoPhys}
\title{Calculate CSR Scores and Assign a CSR Strategy Using the Morphophysiological Model (Novakovskiy et al. 2016)}
\usage{
morphoPhys(data, calcLDMC = FALSE)
}
\arguments{
\item{data}{A data frame containing species names and trait data. Must include (LFW optional in place of LDMC):
\describe{
  \item{CH}{Canopy height (mm)}
  \item{LDMC}{Leaf dry matter content (\%)}
  \item{FP}{Flowering period (months)}
  \item{LS}{Lateral spread (categorical scale: 1–6, based on Hodgson et al. 1999)}
  \item{LDW}{Leaf dry weight (mg)}
  \item{LFW}{Leaf fresh weight(mg)}
  \item{PN}{Net photosynthesis (mg CO₂/g dry weight per hour)}
  \item{RD}{Dark respiration rate (same units as PN)}
  \item{LNC}{Leaf nitrogen content (mg/g)}
  \item{LCC}{Leaf carbon concentration (mg/g)}
}}

\item{calcLDMC}{Logical. Whether to calculate LDMC from LDW and LFW. Default is FALSE.}
}
\value{
A data frame with transformed traits, CSR percentages (C%, S%, R%), strategy classification and optional trait calculations.
}
\description{
This function calculates and assigns CSR strategy types to plant species
based on the morphophysiological CSR model by Novakovskiy et al. (2016), as well as performing
the required trait transformations for the model.
}
\examples{
# morphoPhys(exampleData)
}
